requirements.txt
pandas>=1.5.0
numpy>=1.25.0
openpyxl>=3.1.0
matplotlib>=3.8.0
seaborn>=0.12.2
reportlab>=3.7.0
python-docx>=0.8.12
python-pptx>=0.6.21
